WebConfig File
Running both the REST server and client through Intellij caused CORS errors (it was a web security error which stopped there being both a client and a server coming from the same machine). This required both server and client changes. The angular config file was updated to include 2 config settings with $httpProvider. The REST environment file was added, called 'WebConfig'.

WebConfig file contains code that allows for client and server to work form the same machine.