Testing -
Tried for a number of days to get both (or even either) Unit and Integration Testing up and running on the back-end.

Took a number of unsuccessful approaches at both.

The initial approaches I believed the combination of the annotations at the beginning of each of the test files and tried numerous configurations, later I came to the conclusion that it was configured properly to test in advance and seeing as I couldn't achieve that soley through the @Before/@BeforeEach methods, that an embedded database needed to be used to achieve what I wanted. I went off them to attempt that. The last of the approaches was based around an embedded H2 database to test around (for integration testing)

I believe the heart of the problems that I was having here was

With time constraints, left this testing aspect to complete the rest of the project.

The files of this attempt are scattered...

There is a Broken Tests directory where a few of the attempts I had been attempting are stored.

Within 'Resources' there are a number of sql files that I had been attempting to use alongside the H2 embedded datasource.

Also within 'Resources' was the applicationContext.xml file. This was specifying the h2 embedded-database and for it to use the schema.sql and data.sql scripts to use alongside it (hopefully to simulate the database and whether the logic was correct)

In the 'test' dir there a few attempts at testing controllers and repositories.

Changes made to the pom were extra dependencies such as mockito for mocking.