Login Controller Added
Which includes REST paths for Login, Logout and Register.

Main Controller

MySQL treatment of time is down to seconds, the TimeStamp treatment of time is down to milliseconds, this causes comparison problems with time, as a result I have chosen to eliminate milliseconds in the recording of time. A function was added that is available to all controllers which does this for me (this is called getTimeStamp()).

Brought a method getTimeStamp() to Main Controller, available to the rest of the controllers as they all extend the Main Controller.