H2 Embedded Database removed
The REST side was broken complaining about 'datasource' errors, I concluded that it had to be down to the fact that I had attempted to bring in an embedded H2 database. So I gathered up all the files and dependencies that I could on that front and deleted them. REST worked as expected again following.