JSON References Error
When reading StockItems, no StockReviews were populated in each object
although StockItems were present in StockReview Object. This was caused
by incorrect Jackson Annotations (for bidirectional relationships).
@JsonManagedReference and @JsonBackReference (they were backwards).

To change this, 3 entities had to be changed (all entities in the Many-
to-Many relationship). The 3 entities are Account, StockReview and
StockItem.

Similarly, for StockItem, Manufacturer and ItemCategory. Same error and
changes applied here. These changes were all in the models.

Controllers (@RequestMapping Changes)

Allow the REST to call by StockItem and StockItems in the controller.

@RequestMapping(value = {"StockItems", "StockItem"), method=RequestMethod.GET)

Similar was done for each of the controllers.