Persisting Stock Updates (StockItemController) & Main Controller Update
new method called setNewStockLevel and it responds to the following
mapping "/StockItem/{stockItemId}/{newStockQty}"

This increases/decreases stock levels by stockItemId, using the changes
made to the table used by Administrator. (This will work for customers
as well).

MainController Update

HttpStatus is used by a number of controllers to return status back to
the user, reduce duplication, this was refactored to the main controller
and set to a default value "HttpStatus.OK"